+++ wallet/internal/db/itest/account_store_test.go
@@ -3,12 +3,8 @@
 package itest
 
 import (
-	"context"
 	"fmt"
 	"slices"
-	"sort"
-	"strconv"
-	"sync"
 	"testing"
 	"time"
 
@@ -20,250 +16,6 @@ import (
 
 
 
-// TestCreateDerivedAccountErrors verifies that CreateDerivedAccount returns
-// appropriate errors for invalid inputs.
-func TestCreateDerivedAccountErrors(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		params  db.CreateDerivedAccountParams
-		wantErr error
-	}{
-		{
-			name: "missing name",
-			params: db.CreateDerivedAccountParams{
-				Scope: db.KeyScopeBIP0084,
-				Name:  "",
-			},
-			wantErr: db.ErrMissingAccountName,
-		},
-		{
-			name: "unknown scope",
-			params: db.CreateDerivedAccountParams{
-				Scope: db.KeyScope{Purpose: 999, Coin: 999},
-				Name:  "unknown-scope-account",
-			},
-			wantErr: db.ErrUnknownKeyScope,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			t.Parallel()
-
-			store := NewTestStore(t)
-			walletID := newWallet(t, store, tc.name+"-wallet")
-			tc.params.WalletID = walletID
-
-			info, err := store.CreateDerivedAccount(
-				t.Context(), tc.params, SpendableDeriveFn(),
-			)
-			require.ErrorIs(t, err, tc.wantErr)
-			require.Nil(t, info)
-		})
-	}
-}
-
-// TestCreateDerivedAccountMissingWallet verifies that CreateDerivedAccount
-// returns ErrWalletNotFound when the wallet does not exist.
-func TestCreateDerivedAccountMissingWallet(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	params := db.CreateDerivedAccountParams{
-		WalletID: 99999,
-		Scope:    db.KeyScopeBIP0084,
-		Name:     "missing-wallet-account",
-	}
-
-	info, err := store.CreateDerivedAccount(
-		t.Context(), params, SpendableDeriveFn(),
-	)
-	require.ErrorIs(t, err, db.ErrWalletNotFound)
-	require.Nil(t, info)
-}
-
-// TestCreateDerivedAccountDuplicateName verifies that creating a derived
-// account with a duplicate name in the same scope fails.
-func TestCreateDerivedAccountDuplicateName(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	walletID := newWallet(t, store, "duplicate-name-wallet")
-
-	params := db.CreateDerivedAccountParams{
-		WalletID: walletID,
-		Scope:    db.KeyScopeBIP0084,
-		Name:     "duplicate-account",
-	}
-
-	_, err := store.CreateDerivedAccount(
-		t.Context(), params, SpendableDeriveFn(),
-	)
-	require.NoError(t, err)
-
-	before := store.StatsSnapshot()
-
-	// Attempt to create second account with same name in same scope.
-	_, err = store.CreateDerivedAccount(
-		t.Context(), params, SpendableDeriveFn(),
-	)
-	require.Error(t, err)
-	requireConstraintSQLError(t, err)
-
-	after := store.StatsSnapshot()
-	require.Equal(t, before.Unhealthy, after.Unhealthy)
-	require.Equal(t, before.RetryAttempts, after.RetryAttempts)
-	require.Equal(t, before.RetrySuccesses, after.RetrySuccesses)
-	require.Equal(t, before.RetryExhausted, after.RetryExhausted)
-	require.Equal(t, before.AmbiguousTxCommits, after.AmbiguousTxCommits)
-	require.Equal(t, before.Errors.Backend, after.Errors.Backend)
-	require.Equal(t, before.Errors.TotalErrs+1, after.Errors.TotalErrs)
-	require.Equal(
-		t,
-		before.Errors.PermanentErrs+1,
-		after.Errors.PermanentErrs,
-	)
-	require.Equal(t, before.Errors.Constraint+1, after.Errors.Constraint)
-	require.Equal(t, before.Errors.TransientErrs, after.Errors.TransientErrs)
-	require.Equal(t, before.Errors.FatalErrs, after.Errors.FatalErrs)
-}
-
-// TestCreateDerivedAccountSameNameDifferentScopes verifies that accounts with
-// the same name can exist in different scopes.
-func TestCreateDerivedAccountSameNameDifferentScopes(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	walletID := newWallet(t, store, "multi-scope-wallet")
-
-	accountName := "shared-name"
-	scopes := []db.KeyScope{
-		db.KeyScopeBIP0084,
-		db.KeyScopeBIP0086,
-		db.KeyScopeBIP0044,
-		db.KeyScopeBIP0049Plus,
-	}
-
-	for _, scope := range scopes {
-		params := db.CreateDerivedAccountParams{
-			WalletID: walletID,
-			Scope:    scope,
-			Name:     accountName,
-		}
-
-		info, err := store.CreateDerivedAccount(
-			t.Context(), params, SpendableDeriveFn(),
-		)
-		require.NoError(t, err)
-		require.Equal(t, accountName, info.AccountName)
-		require.Equal(t, scope, info.KeyScope)
-	}
-}
-
-// TestCreateDerivedAccountSequentialNumbers verifies that derived accounts
-// within the same scope receive sequential account numbers starting from 0.
-func TestCreateDerivedAccountSequentialNumbers(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	walletID := newWallet(t, store, "sequential-wallet")
-
-	scope := db.KeyScopeBIP0084
-
-	const numAccounts = 5
-
-	for i := range numAccounts {
-		params := db.CreateDerivedAccountParams{
-			WalletID: walletID,
-			Scope:    scope,
-			Name:     "account-" + strconv.Itoa(i),
-		}
-
-		info, err := store.CreateDerivedAccount(
-			t.Context(), params, SpendableDeriveFn(),
-		)
-		require.NoError(t, err)
-		require.Equal(t, uint32(i), info.AccountNumber,
-			"account %d should have number %d", i, i)
-	}
-}
-
-// TestCreateDerivedAccountConcurrent verifies that concurrent account creation
-// yields unique, sequential account numbers without errors.
-func TestCreateDerivedAccountConcurrent(t *testing.T) {
-	t.Parallel()
-
-	store := NewTestStore(t)
-
-	walletID := newWallet(t, store, "concurrent-wallet")
-
-	scope := db.KeyScopeBIP0084
-
-	const workers = 20
-
-	type createResult struct {
-		number uint32
-		err    error
-	}
-
-	resultCh := make(chan createResult, workers)
-
-	var wg sync.WaitGroup
-
-	ctx, cancel := context.WithTimeout(t.Context(), 5*time.Second)
-	defer cancel()
-
-	for i := range workers {
-		wg.Add(1)
-
-		go func(i int) {
-			defer wg.Done()
-
-			info, err := store.CreateDerivedAccount(
-				ctx, db.CreateDerivedAccountParams{
-					WalletID: walletID,
-					Scope:    scope,
-					Name:     "acct-concurrent-" + strconv.Itoa(i),
-				},
-				SpendableDeriveFn(),
-			)
-			if err != nil {
-				resultCh <- createResult{err: err}
-				return
-			}
-
-			resultCh <- createResult{number: info.AccountNumber}
-		}(i)
-	}
-
-	wg.Wait()
-	close(resultCh)
-
-	results := make([]uint32, 0, workers)
-	for result := range resultCh {
-		require.NoError(t, result.err)
-		results = append(results, result.number)
-	}
-
-	require.Len(t, results, workers)
-
-	// Verify all numbers are unique and sequential.
-	sort.Slice(
-		results, func(i, j int) bool {
-			return results[i] < results[j]
-		},
-	)
-
-	for i := range workers {
-		require.Equal(t, uint32(i), results[i])
-	}
-}
 
 // TestCreateImportedAccountErrors verifies that CreateImportedAccount returns
 // appropriate errors for invalid inputs.
